+++ example_app/app.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from flask_pydantic.openapi import APIError
 from typing import Optional
 
-from flask import Flask, request, jsonify
+from flask import Flask, jsonify, request
 from flask_pydantic import validate, add_openapi_spec, openapi_docs
 from pydantic import BaseModel
 
@@ -26,6 +26,11 @@ class IndexParam(BaseModel):
 
 
 class BodyModel(BaseModel):
+    name: str
+    nickname: Optional[str]
+
+
+class FormModel(BaseModel):
     name: str
     nickname: Optional[str]
 
@@ -55,6 +60,23 @@ def post():
     )
 
 
+@app.route("/form", methods=["POST"])
+@validate(form=FormModel, query=QueryModel)
+def post():
+    """
+    Basic example with both query and form-data parameters, response object serialization.
+    """
+    # save model to DB
+    id_ = 2
+
+    return ResponseModel(
+        id=id_,
+        age=request.query_params.age,
+        name=request.form_params.name,
+        nickname=request.form_params.nickname,
+    )
+
+
 @app.route("/kwargs", methods=["POST"])
 @openapi_docs(response=ResponseModel, tags=["kwargs"])
 @validate()
@@ -67,6 +89,19 @@ def post_kwargs(body: BodyModel, query: QueryModel):
     id_ = 3
 
     return ResponseModel(id=id_, age=query.age, name=body.name, nickname=body.nickname)
+
+
+@app.route("/form/kwargs", methods=["POST"])
+@validate()
+def post_kwargs(form: FormModel, query: QueryModel):
+    """
+    Basic example with both query and form-data parameters, response object serialization.
+    This time using the decorated function kwargs `form` and `query` type hinting
+    """
+    # save model to DB
+    id_ = 3
+
+    return ResponseModel(id=id_, age=query.age, name=form.name, nickname=form.nickname)
 
 
 @app.route("/many", methods=["GET"])